add abstraction of repositories at the interface level
