+++ src/main/java/ru/yandex/practicum/filmorate/repository/UserRepository.java
@@ -1,33 +1,15 @@
 package ru.yandex.practicum.filmorate.repository;
 
-import org.springframework.stereotype.Repository;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
 
-@Repository
-public class UserRepository {
-    private static long identifierUser = 1;
-    private final Map<Long, User> users = new HashMap<>();
+public interface UserRepository {
+    User create(User user);
 
-    public User create(User user) {
-        user.setId(identifierUser++);
-        users.put(user.getId(), user);
-        return user;
-    }
+    User update(User user);
 
-    public User update(User user) {
-        users.put(user.getId(), user);
-        return user;
-    }
+    Collection<User> getAll();
 
-    public Collection<User> getAll() {
-        return users.values();
-    }
-
-    public User getById(Long id) {
-        return users.get(id);
-    }
+    User getById(Long id);
 }
